+++ KnowYourInterview/api/src/main/java/com/knowyourinterview/api/payment/dto/CreateOrderResponse.java
@@ -3,5 +3,5 @@ package com.knowyourinterview.api.payment.dto;
 import java.util.UUID;
 
 public record CreateOrderResponse(
-        UUID experienceId, String razorpayOrderId, int amountPaise, String currency, String razorpayKeyId) {
+        UUID experienceId, String razorpayOrderId, long amountPaise, String currency, String razorpayKeyId) {
 }
